OPH-921 | Refactor breadcrums to show main route (#179)
* refactor: shared-processes breadcrums show the shared-processes as breadcrum instead of all processes

Refs: oph-921

* refactor: remove new route and controller test

Refs: oph-921

* refactor: also show the correct parent breadcrum when in local gov processes route

Refs: oph-921

* refactor: remove the index route from the functions in the router

Refs: oph-921

* refactor: just use the transition.parent.name for the modelFor method

Refs: oph-921

* refactor: remove all process/index files as it is not needed

Refs: oph-921

* fix: include relations again when fetching the process

Refs: oph-921
