+++ app/controllers/my-local-government/process.js
@@ -1,0 +1,3 @@
+import ProcessesProcessController from '../processes/process';
+
+export default class MyLocalGovernmentProcessController extends ProcessesProcessController {}
