+++ app/controllers/shared-processes/process.js
@@ -1,0 +1,3 @@
+import ProcessesProcessController from '../processes/process';
+
+export default class SharedProcessesProcessController extends ProcessesProcessController {}
