+++ app/router.js
@@ -28,30 +28,26 @@ Router.map(function () {
   this.route('index', { path: '' });
 
   this.route('processes', { path: 'processen' }, function () {
-    this.route('process', { path: '/:id/' }, function () {
-      this.route('index', { path: '' });
-    });
+    this.route('process', { path: '/:id/' }, function () {});
   });
 
   this.route('shared-processes', { path: 'gedeelde-processen' }, function () {
-    this.route('index', { path: '' });
+    this.route('process', { path: '/:id/' }, function () {});
   });
 
   this.route(
     'my-local-government',
     { path: 'mijn-lokaal-bestuur' },
     function () {
-      this.route('index', { path: '' });
+      this.route('process', { path: '/:id/' }, function () {});
     },
   );
 
   this.route('inventory', { path: 'inventaris' }, function () {
-    this.route('index', { path: '' });
     this.route('admin');
   });
 
   this.route('information-assets', { path: 'informatie-assets' }, function () {
-    this.route('index', { path: '' });
     this.route('edit', { path: '/:id' });
   });
