+++ app/routes/my-local-government/process.js
@@ -1,0 +1,3 @@
+import ProcessesProcessRoute from '../processes/process';
+
+export default class MyLocalGovernmentProcessRoute extends ProcessesProcessRoute {}
