+++ app/routes/processes/process.js
@@ -2,9 +2,8 @@ import Route from '@ember/routing/route';
 
 import { service } from '@ember/service';
 
-import ENV from 'frontend-openproceshuis/config/environment';
-
 export default class ProcessesProcessRoute extends Route {
+  @service plausible;
   @service store;
   @service router;
   @service session;
@@ -15,30 +14,61 @@ export default class ProcessesProcessRoute extends Route {
     }
   }
 
-  async model(params) {
-    try {
-      const process = await this.store.findRecord('process', params.id, {
-        'filter[files][:not:status]': ENV.resourceStates.archived,
-        include: [
-          'process-statistics',
-          'files',
-          'publisher',
-          'publisher.primary-site',
-          'publisher.primary-site.contacts',
-          'publisher.classification',
-          'ipdc-products',
-          'information-assets',
-          'linked-concept',
-          'linked-concept.process-groups.process-domains',
-          'linked-concept.process-groups.process-domains.process-categories',
-          'relevant-administrative-units',
-        ].join(','),
-        reload: true,
-      });
+  async model({ id }) {
+    const process = await this.store.findRecord('process', id, {
+      include: [
+        'process-statistics',
+        'files',
+        'publisher',
+        'publisher.primary-site',
+        'publisher.primary-site.contacts',
+        'publisher.classification',
+        'ipdc-products',
+        'information-assets',
+        'linked-concept',
+        'linked-concept.process-groups.process-domains',
+        'linked-concept.process-groups.process-domains.process-categories',
+        'relevant-administrative-units',
+      ].join(','),
+    });
+    let stats = process.processStatistics;
 
-      return process;
-    } catch (error) {
-      throw new Error(`Er werd geen process met id "${params.id}" gevonden.`);
+    if (!stats) {
+      stats = this.store.createRecord('process-statistic', {
+        process,
+      });
+    }
+
+    this.plausible.trackEvent('Raadpleeg proces', {
+      'Proces-ID': process?.id,
+      Procesnaam: process?.title,
+      'Bestuur-ID': process?.publisher?.id,
+      Bestuursnaam: process?.publisher?.name,
+    });
+
+    stats.processViews += 1;
+    await stats.save();
+
+    return {
+      process,
+      breadcrumRouteName: this.router.currentRouteName?.replace('.index'),
+    };
+  }
+
+  setupController(controller) {
+    super.setupController(...arguments);
+    controller.reset();
+  }
+
+  resetController(controller) {
+    super.resetController(...arguments);
+    controller.reset();
+  }
+  didTransition() {
+    const scrollTo = this.paramsFor('processes.process')?.scrollTo;
+    if (scrollTo) {
+      const el = document.getElementById(scrollTo);
+      if (el) el.scrollIntoView({ behavior: 'smooth', block: 'start' });
     }
   }
 }
