+++ app/routes/shared-processes/process.js
@@ -1,0 +1,3 @@
+import ProcessesProcessRoute from '../processes/process';
+
+export default class SharedProcessesProcessRoute extends ProcessesProcessRoute {}
