+++ app/templates/my-local-government/index.hbs
@@ -97,7 +97,7 @@
               <LinkTo
                 class="au-c-link"
                 @model={{process.id}}
-                @route="processes.process"
+                @route="my-local-government.process"
               >
                 {{process.title}}
               </LinkTo>
