+++ app/templates/my-local-government/process.js
@@ -1,0 +1,1 @@
+export { default } from '../processes/process';
