--- app/templates/processes/process/index.hbs
@@ -1,76 +1,0 @@
-{{#if this.isEditingDetails}}
-  <PageHeader>
-    <:title>
-      Wijzig proces
-    </:title>
-    <:action>
-      <AuPill @icon="clock-rewind" @draft="true">Laatst aangepast op
-        {{date-format this.process.modified true}}</AuPill>
-    </:action>
-  </PageHeader>
-{{else}}
-  <PageHeader>
-    <:title>
-      <div class="au-u-flex au-u-flex--vertical-center">
-        {{this.process.title}}
-        <AuButton
-          class="au-u-padding-left-small"
-          @icon="link"
-          @skin="link"
-          @hideText={{true}}
-          {{on "click" this.copyUrl}}
-        /></div>
-    </:title>
-    <:action>
-      <AuPill @icon="clock-rewind">Laatst aangepast op
-        {{date-format this.process.modified true}}</AuPill>
-    </:action>
-  </PageHeader>
-{{/if}}
-
-<AuBodyContainer class="au-o-box au-o-flow" @scroll={{true}}>
-
-  <Process::DetailsCard
-    @canEdit={{this.canEdit}}
-    @process={{this.process}}
-    @setIsEditing={{this.setIsEditingDetails}}
-  />
-
-  <Process::InventoryProcessCard
-    @canEdit={{this.canEdit}}
-    @process={{this.process}}
-  />
-
-  <Process::Diagram::Visual
-    @canEdit={{this.canEdit}}
-    @process={{this.process}}
-    @trackDownloadFileEvent={{this.trackDownloadFileEvent}}
-  />
-
-  {{#if (and this.process.isBlueprint this.process.isPublishedByAbbOrDv)}}
-    <Process::IcrCard @canEdit={{this.canEdit}} @process={{this.process}} />
-  {{else}}
-    <Process::BlueprintCard
-      @canEdit={{this.canEdit}}
-      @process={{this.process}}
-    />
-  {{/if}}
-
-  <Process::Diagram::Version
-    @canEdit={{this.canEdit}}
-    @process={{this.process}}
-    @trackDownloadFileEvent={{this.trackDownloadFileEvent}}
-  />
-
-  <Process::Attachments
-    @canEdit={{this.canEdit}}
-    @process={{this.process}}
-    @trackDownloadFileEvent={{this.trackDownloadFileEvent}}
-  />
-
-  <Process::Usage
-    @process={{this.process}}
-    @refreshDataTrigger={{this.isEditingDetails}}
-  />
-
-</AuBodyContainer>
