+++ app/templates/shared-processes/index.hbs
@@ -94,7 +94,7 @@
               <LinkTo
                 class="au-c-link"
                 @model={{process.id}}
-                @route="processes.process"
+                @route="shared-processes.process"
               >
                 {{process.title}}
               </LinkTo>
