+++ app/templates/shared-processes/process.js
@@ -1,0 +1,1 @@
+export { default } from '../processes/process';
